move BeneficiaryFilterType to endpoints file
