+++ src/database/models/ubi/beneficiary.ts
@@ -3,14 +3,6 @@ import { Sequelize, DataTypes, Model } from 'sequelize';
 
 import { CommunityAttributes } from './community';
 
-export type BeneficiaryFilterType = {
-    active?: boolean;
-    suspect?: boolean;
-    inactivity?: boolean;
-    unidentified?: boolean;
-    blocked?: boolean;
-};
-
 export interface BeneficiaryAttributes {
     id: number;
     address: string;
